Fix error filtering
