+++ src/components/UI/JSONSchemaForm/ErrorListTemplate/index.tsx
@@ -1,24 +1,13 @@
 import { ErrorListProps, RJSFSchema } from '@rjsf/utils';
-import React, { useMemo } from 'react';
+import React from 'react';
 
 import { IFormContext } from '..';
-import { isTouchedField, mapErrorPropertyToField } from '../utils';
-import ErrorList from './ErrorList';
+import ValidationStatus from './ValidationStatus';
 
 const ErrorListTemplate: React.FC<
   ErrorListProps<RJSFSchema, RJSFSchema, IFormContext>
-> = ({ errors, formContext = {} as IFormContext }) => {
-  const filteredErrors = useMemo(() => {
-    return errors.filter((e) =>
-      isTouchedField(
-        mapErrorPropertyToField(e, formContext.idConfigs),
-        formContext.touchedFields,
-        formContext.idConfigs.idSeparator
-      )
-    );
-  }, [errors, formContext]);
-
-  return <ErrorList errors={filteredErrors} />;
+> = ({ formContext = {} as IFormContext }) => {
+  return <ValidationStatus formContext={formContext} />;
 };
 
 export default ErrorListTemplate;
